pyproject.toml
[project]
name = "tcc-questions"
version = "0.1.0"
description = "Sistema de catalogação de respostas LLM para TCC"
authors = [
    {name = "Bruno", email = "bruno@example.com"}
]
dependencies = [
    "fastapi[standard]>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.23",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.12.1",
    "jinja2>=3.1.2",
    "python-multipart>=0.0.6",
    "pandas>=2.1.3",
    "numpy>=1.25.2",
    "scipy>=1.11.4",
    "plotly>=5.17.0",
    "python-dotenv>=1.0.0",
    "openai>=1.3.0",
    "groq>=0.4.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "instructor>=1.9.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88